+++ parquet/src/util/test_common/mod.rs
@@ -22,3 +22,6 @@ pub mod file_util;
 
 #[cfg(test)]
 pub mod rand_gen;
+
+#[cfg(all(test, feature = "encryption", feature = "arrow"))]
+pub mod encryption_util;
